+++ sync-for-payables/src/codatsyncpayables/manage_data.py
@@ -113,7 +113,13 @@ class ManageData:
     
     def list_pull_operations(self, request: operations.ListPullOperationsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListPullOperationsResponse:
         r"""List pull operations
-        Gets the pull operation history (datasets) for a given company.
+        The *List pull operations* endpoint returns a list of [pull operations](https://docs.codat.io/sync-for-payables-api#/schemas/PullOperation) made by your client.
+
+        A [pull operation](https://docs.codat.io/sync-for-payables-api#/schemas/PullOperation) is a request to retrieve a specific data type from an integration.
+
+        ### Tips and traps
+
+        - The *List pull operations* endpoint does not support querying the `isCompleted` property. You can filter failed pull operations by querying `status!=Complete&&status!=NotSupported` instead.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())